Better version handling, consistent logs, bugs
We now check the installed package for the current version. This allows
us to move to omnibus, even if the current chef version (e.g. of the
gem) is the same as the one to be installed.

Also, we now check the full version including the build-version if the
download URL is provided. Unfortunately, the Omnitruck API doesn't allow
us to specify this. Thus, if you need a new build version, you have to
enter the full download URL. If only the target version is specified, we
only check the minor version and don't check with Omnitruck if it matches.

We now provide better debug logs which are more consistent with the logs
provided by the package resource.

Finally, we fixed some remaining bugs, esp. with regard to Chef 10
compatibility.